A* heuristic based on distance to the goal (#3)
Compute the distance to the goal using the current state value, instead of a simple incrementation using the number of steps from the starting point.
Heap storage for the nodes, to reduce the number of slices & maps. Reduces the fetch of next node.
One call to fetchNodeInHeap instead of two.
Hash with XOR incrementation.
Simplified hash using a simple hash multiplication.
Unit Test coverage > 90%.
Code documentation.